requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filter==23.3
django-webpack-loader==1.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
install==1.3.5
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.2
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.0
protobuf==4.21.12
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
